Fix bug with key calculation
Key is calculated in a loop and so may be overflowed. Since int may be
different size on different platforms, calculated key may be different
too and incompatible on some platforms. Setting size explicitly 32 bit
removes that UB (code will drop support for some platforms though, but
that's not that important to think of)